#105 Move type hint of reference
