+++ xarray_dataclasses/dataarray.py
@@ -16,8 +16,7 @@ from typing_extensions import Literal, ParamSpec, Protocol
 
 
 # submodules
-from .datamodel import DataModel
-from .typing import Reference
+from .datamodel import DataModel, Reference
 
 
 # type hints
