+++ xarray_dataclasses/typing.py
@@ -20,7 +20,6 @@ from typing import Any, Dict, Optional, Tuple, TypeVar, Union
 
 
 # dependencies
-import xarray as xr
 from typing_extensions import (
     Annotated,
     get_args,
@@ -66,7 +65,6 @@ TDtype = TypeVar("TDtype", covariant=True)
 Dims = Tuple[str, ...]
 Dtype = Optional[str]
 NoneType = type(None)
-Reference = Union[xr.DataArray, xr.Dataset, None]
 
 
 @runtime_checkable
